Project 2 for Advanced C++ Course
The program prompts the user for the time of the observation, as well as
the temperature. The program accepts the time in the format HH:MM, and
rejects any time less than 00:00, and greater than 23:59. In the case of
an invalid time entry, "Time is out of range" is printed to STDERR and
program execution continues.

Your program should accept no more than 24 observations. If over 24
observations are entered, the program shall return "Maximum number of
observations recorded" to STDERR and immediately show the results
(average, highs and lows).