Project 5 for Advanced C++ Course
The program has SAE.cpp and SAE.h. It has implemented the following
operators: +, +=, -, -=, *, *=, /, /=, and <<. It also has two versions
of the multiply and divide functions: one that takes an SAE argument,
and one that takes a scalar (integer, float, double, etc). It uses
templates for the scalar version of the function. For the multiply and
divide function that do not take scalars,  increases or decreases the
dimension appropriately.
The << operator is smart enough about the output. It shows feet and
inches for 1 dimensional values, and for multidimensional values, it
prints the value in inches only if the value is less than could be
represented with the smallest foot unit.