Project 1 for Advanced C++ Course
First part: This program is a fire and control computer for the Army.
The initial design implements a program that takes two inputs: the angle
of the gun (in degrees) and the velocity of the projectile (in meters
per second), and returns the distance the projectile traveled. The
result is printed in both meters and feet.

Second part: This program is a helper for a manufacturer of printed
fabrics. These fabrics can be used on everything from handkerchiefs to
bed spreads. Depending on the intended final product, the print is
scaled so the entire design can fit on the fabric. This design is the
pine tree. The program prints templates for trees of any size.